requirements.txt
alembic==0.8.10
aniso8601==1.2.0
appdirs==1.4.0
click==6.7
dominate==2.3.1
Flask==0.12
Flask-KVSession==0.6.2
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Moment==0.5.1
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
packaging==16.8
pipdeptree==0.9.0
psycopg2==2.6.2
pyparsing==2.1.10
python-dateutil==2.6.0
python-dotenv==0.6.3
python-editor==1.0.3
pytz==2016.10
redis==2.10.5
requests==2.13.0
simplekv==0.10.0
six==1.10.0
SQLAlchemy==1.1.5
virtualenv==15.1.0
visitor==0.1.3
Werkzeug==0.11.15
WTForms==2.1
